+++ HISTORY.md
@@ -3,6 +3,19 @@
 History
 -------
 
+0.2.0 (2016-07-11)
+------------------
+
+* webui calls for configuration
+
+
+0.1.1 (2016-03-17)
+------------------
+
+* fix value for dynamic default args
+
+
+
 0.1.0 (2015-12-10)
 ---------------------